[feature/PEER-11-AppreciationCRUD] review comments resolved

+++ internal/pkg/dto/appreciation.go
@@ -8,10 +8,10 @@ import (
 
 type Appreciation struct {
 	ID                int64  `json:"id"`
-	CoreValueID       int    `json:"core_value_id" `
+	CoreValueID       int64    `json:"core_value_id" `
 	Description       string `json:"description"`
-	TotalRewardPoints int    `json:"total_reward_points,omitempty"`
-	Quarter           int    `json:"quarter"`
+	TotalRewardPoints int32    `json:"total_reward_points,omitempty"`
+	Quarter           int8    `json:"quarter"`
 	Sender            int64  `json:"sender"`
 	Receiver          int64  `json:"receiver"`
 	CreatedAt         int64  `json:"created_at"`
@@ -27,11 +27,11 @@ type AppreciationFilter struct {
 }
 
 type ResponseAppreciation struct {
-	ID                  int    `json:"id"`
+	ID                  int64    `json:"id"`
 	CoreValueName       string `json:"core_value_name"`
 	CoreValueDesc       string `json:"core_value_description"`
 	Description         string `json:"description"`
-	TotalRewardPoints   int    `json:"total_reward_points"`
+	TotalRewardPoints   int32    `json:"total_reward_points"`
 	Quarter             string `json:"quarter"`
 	SenderFirstName     string `json:"sender_first_name"`
 	SenderLastName      string `json:"sender_last_name"`
@@ -41,8 +41,8 @@ type ResponseAppreciation struct {
 	ReceiverLastName    string `json:"receiver_last_name"`
 	ReceiverImageURL    string `json:"receiver_image_url"`
 	ReceiverDesignation string `json:"receiver_designation"`
-	TotalRewards        int    `json:"total_rewards"`
-	GivenRewardPoint    int    `json:"given_reward_point"`
+	TotalRewards        int32    `json:"total_rewards"`
+	GivenRewardPoint    int8    `json:"given_reward_point"`
 	CreatedAt           int64  `json:"created_at"`
 	UpdatedAt           int64  `json:"updated_at"`
 }
